requirements.txt
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
Backtesting==0.3.3
beautifulsoup4==4.10.0
black==22.1.0
bokeh==2.4.2
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
chromedriver-py==99.0.4844.35
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
DateTime==4.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.2
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-filter==21.1
django-inline-svg==0.1.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
et-xmlfile==1.1.0
executing==0.8.2
finance-datareader==0.9.31
gunicorn==20.1.0
h11==0.13.0
idna==3.3
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pymongo==4.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
qm @ git+https://github.com/jungwon-moon/qm@3f3173447fc7259c9bf367a3d5944cd9685103d3
requests==2.27.1
requests-file==1.5.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.2.4
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
zope.interface==5.4.0
uwsgi